htt-application-master/python-prototypes/ImageAcquisition/requirements.txt
altgraph==0.17
appdirs==1.4.4
astroid==2.4.1
colorama==0.4.3
configparser==5.0.0
distlib==0.3.0
filelock==3.0.12
future==0.18.2
imageio==2.8.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.18.5
pefile==2019.4.18
Pillow==7.1.2
psutil==5.7.0
pybind11==2.5.0
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.7
pylint==2.5.2
PyQt5==5.15.0
PyQt5-sip==12.8.0
pyserial==3.4
python-config==0.1.2
pywin32==228
pywin32-ctypes==0.2.0
six==1.14.0
toml==0.10.1
virtualenv==20.0.20
wrapt==1.12.1
